mcp: carve tools_managed.rs (pure move)
The managed-context and controller-lifecycle tool implementations move
into their own impl block (inherent impls may split across modules):
rewind_context/anchors/backout, fission spawn/control/claim + the
fission wait/cancel free helpers, and the controller
restart/halt/intervention/status tools — with their 47 tests (2,391
lines) and the fission test helpers. The #[tool] attrs on these methods
are manual-dispatch metadata (this impl block has no #[tool_router]);
moving them changes no router registration.

Usual glue + manufactured `impl IntendantServer {}` wrapper;
purity-checked; mcp:: tests 178/178.
